scripts/requirements-plant-disease.txt
numpy>=1.26
Pillow>=10.0
tflite-runtime>=2.14
